Update TinyMCE to 2017-01-03 nightly build. Fix #8
I submitted an issue to the TinyMCE repository and got them to implement my desired functionality for FontSizeSelect as of the 2017-01-03 nightly build. This closes #8, but beware of other bugs as the nightly builds are often unstable. TinyMCE is also no longer being served from a CDN and is instead stored locally at static/js/tinymce.